+++ lib/pane/input-point-3d-test.ts
@@ -6,8 +6,8 @@ import {TestUtil} from '../misc/test-util';
 import {findConstraint} from '../plugin/common/constraint/composite';
 import {RangeConstraint} from '../plugin/common/constraint/range';
 import {StepConstraint} from '../plugin/common/constraint/step';
-import {PointNdConstraint} from '../plugin/input-bindings/point-2d/constraint/point-nd';
-import {PointNdTextController} from '../plugin/input-bindings/point-2d/controller/point-nd-text';
+import {PointNdConstraint} from '../plugin/input-bindings/common/constraint/point-nd';
+import {PointNdTextController} from '../plugin/input-bindings/common/controller/point-nd-text';
 
 function createPane(): Tweakpane {
 	return new Tweakpane({
