+++ lib/plugin/input-bindings/common/constraint/point-nd.ts
