+++ lib/plugin/input-bindings/common/controller/point-nd-text.ts
