+++ lib/plugin/input-bindings/common/model/point-nd.ts
