+++ lib/plugin/input-bindings/common/view/point-nd-text.ts
