+++ lib/plugin/input-bindings/point-2d/controller/point-2d-pad-text.ts
@@ -3,10 +3,10 @@ import {ValueController} from '../../../common/controller/value';
 import {Formatter} from '../../../common/converter/formatter';
 import {Parser} from '../../../common/converter/parser';
 import {Value} from '../../../common/model/value';
+import {PointNdTextController} from '../../common/controller/point-nd-text';
 import {Point2d, Point2dAssembly} from '../model/point-2d';
 import {Point2dPadTextView} from '../view/point-2d-pad-text';
 import {Point2dPadController} from './point-2d-pad';
-import {PointNdTextController} from './point-nd-text';
 
 interface Axis {
 	baseStep: number;
