+++ lib/plugin/input-bindings/point-2d/plugin.ts
@@ -19,7 +19,7 @@ import {
 	getSuitableDecimalDigits,
 	getSuitableDraggingScale,
 } from '../../util';
-import {PointNdConstraint} from './constraint/point-nd';
+import {PointNdConstraint} from '../common/constraint/point-nd';
 import {Point2dPadTextController} from './controller/point-2d-pad-text';
 import {point2dFromUnknown, writePoint2d} from './converter/point-2d';
 import {Point2d, Point2dAssembly, Point2dObject} from './model/point-2d';
